♻️ [refactor]: #452 Phase 4: lexer_skip_whitespace.rs を分離
skip_whitespace の 7 テストを `src/lexer/tests/lexer_skip_whitespace.rs` に分離。

- 293 件グリーン維持
- 短縮テスト関数名集合不変 / 重複 0
- lexer_skip_whitespace.rs: 64 行 (< 800)

Refs #452
